requirements.txt
fastapi
pymupdf
regex
python-dotenv
python-multipart
uvicorn
sqlalchemy[asyncio]
pymysql